Reference

OptimalTrainControl.DavisResistance

OptimalTrainControl.FlatTrack

OptimalTrainControl.FlatTrack

OptimalTrainControl.HillyTrack

OptimalTrainControl.HillyTrack

OptimalTrainControl.Segment

OptimalTrainControl.TrainProblem

OptimalTrainControl.calculatecontrol!

OptimalTrainControl.finish

OptimalTrainControl.finish

OptimalTrainControl.getgrade

OptimalTrainControl.getgrade

OptimalTrainControl.getgradientacceleration

OptimalTrainControl.modecolor

OptimalTrainControl.modecolor

OptimalTrainControl.resistance

OptimalTrainControl.solve!

OptimalTrainControl.start

OptimalTrainControl.start

OptimalTrainControl.subtrack

OptimalTrainControl.subtrack

OptimalTrainControl.subtrack

OptimalTrainControl.DavisResistance — Type

Empirical formula originally calculated for freight cars. The resistance (in N/kg) is given by R = a + b * v + c * v^2, where v is the vehicle speed.

OptimalTrainControl.FlatTrack — Type

FlatTrack(X::Real)

Create instance of a flat track of length X.

OptimalTrainControl.FlatTrack — Method

my_flattrack(x)

Return 0 height value at the distance x.

OptimalTrainControl.HillyTrack — Method

HillyTrack(Xs, Ys)

Create instance of a track. The track is defined by a set of waypoints along distance Xs and heights Ys at the appropriate x-points.

OptimalTrainControl.HillyTrack — Method

my_track(x)

Return height value at the distance x.

OptimalTrainControl.Segment — Type

Segment(start, finish, mode, holdspeed)

Define segment of a track between start and finish where mode is possible with holding speed holdspeed.

OptimalTrainControl.TrainProblem — Type

TrainProblem(; track::Track, T, umax = v -> 3 / max(5, v), umin = v -> -3 / max(5, v),
    resistance = DavisResistance(1e-2, 0, 1.5e-5), ρ = 0, vᵢ = 1.0, vf = 1.0)

Define the optimal train control problem via specifying all of the relevant parameters as keyword arguments. All of the arguments have default values except track and total journey time T.

Arguments

track::Track: Track specification

T: Total time of journey in seconds

umax = v -> 3 / max(5, v): Maximum traction limit, output in (m/s^2)

umin = v -> -3 / max(5, v): Maximum braking traction limit, output in (m/s^2)

resistance = DavisResistance(1e-2, 0, 1.5e-5): Resistance function, output in (m/s^2)

ρ = 0: Regeneration coefficient, 0 ≤ ρ < 1; tells how much braking energy can be reused

vᵢ = 1.0: Initial speed (m/s)

vf = 1.0: Final speed (m/s)

OptimalTrainControl.calculatecontrol! — Method

calculatecontrol!(prob::TrainProblem, sol, points)

Return vector of control signal values evaluated at the distances given by sol.t.

OptimalTrainControl.finish — Method

finish(t::FlatTrack)

Return length of the flat track t in metres.

OptimalTrainControl.finish — Method

finish(t::HillyTrack)

Return length of the track t in metres.

OptimalTrainControl.getgrade — Method

getgrade(track::HillyTrack, pos)

Get the track grade (in radians) of the track at the given position pos.

OptimalTrainControl.getgrade — Method

Get the grade (in radians) of a flat track at the given position.

OptimalTrainControl.getgradientacceleration — Method

getgradientacceleration(::Track, position)

Return acceleration/decelerating component of the gravitational acceleration g = 9.81 m / s^2. Positive means downhill, negative means uphill.

OptimalTrainControl.modecolor — Method

modecolor(Xs, points)

Return vector of colors at positions Xs colored according to the control modes defined by points.

OptimalTrainControl.modecolor — Method

modecolor(mode ∈ ControlModes)

Return color corresponding to the mode.

OptimalTrainControl.resistance — Method

resistance(r::DavisResistance, v)

Calculate the Davis formula resistant force per unit mass.

R = r.a + r.b * v + r.c * v^2,

where v is the vehicle speed and r.a, v.b and v.c are the resistance parameters.

OptimalTrainControl.solve! — Method

solve!(prob::TrainProblem; atol = 5)

Solve the optimal train control problem prob with the total journey time satisfied with the absolute tolerance atol. Return value is a tuple of optimal switching points and differential equation solution.

Examples

points, sol = solve!(prob)

OptimalTrainControl.start — Method

start(t::FlatTrack)

Return 0 as the beginning of the flat track t.

OptimalTrainControl.start — Method

start(t::HillyTrack)

Return start of the track t in metres.

OptimalTrainControl.subtrack — Method

subtrack(full_flattrack::FlatTrack, from)

Create subtrack of type FlatTrack of full_flattrack from from until finish(full_flattrack).

OptimalTrainControl.subtrack — Method

subtrack(full_track::HillyTrack, from, to)

Create subtrack of type HillyTrack of full_track spanning from from to to.

OptimalTrainControl.subtrack — Method

subtrack(full_track::HillyTrack, from)

Create subtrack of type HillyTrack of full_track from from until finish(full_track).