RailDynamics

Documentation for RailDynamics.

RailDynamics.DavisResistance

RailDynamics.albrecht_odefun!

RailDynamics.finish

RailDynamics.finish

RailDynamics.getgrade

RailDynamics.getgrade

RailDynamics.getgradientacceleration

RailDynamics.getgradientacceleration

RailDynamics.inclinationforce

RailDynamics.resistance

RailDynamics.start

RailDynamics.start

RailDynamics.DavisResistance — Type

Empirical formula originally calculated for freight cars. The resistance (in N/kg) is given by R = a + b * v + c * v^2, where v is the vehicle speed.

RailDynamics.albrecht_odefun! — Method

Give the differential equations given in Albrecht et al. 2016, equations (1) and (2). First state is time (second), second state is speed (metre per second).

RailDynamics.finish — Method

finish(t::FlatTrack)

Return length of the flat track in metres.

RailDynamics.finish — Method

finish(t::HillyTrack)

Return length of the track in metres.

RailDynamics.getgrade — Method

getgrade(track::HillyTrack, pos)

Get the track grade (in radians) of the track at the given position pos.

RailDynamics.getgrade — Method

Get the grade (in radians) of a flat track at the given position.

RailDynamics.getgradientacceleration — Method

getgradientacceleration(::Scenario, position)

Return accelerating/decelerating component of the gravitational acceleration in meters per second squared [m s^-2] based on the gravitational acceleration value defined in the scenario. Positive means downhill, negative means uphill.

RailDynamics.getgradientacceleration — Method

getgradientacceleration(::Track, position)

Return acceleration/decelerating component of the gravitational acceleration g = 9.81 m / s^2. Positive means downhill, negative means uphill.

RailDynamics.inclinationforce — Method

Calculate the acting force component based on the current track grade. The returned value is in N/kg, i.e. force per unit mass.

RailDynamics.resistance — Method

Calculate the Davis formula resistant force per unit mass.

R = a + b * v + c * v^2,

where v is the vehicle speed and a, b and c are the resistance parameters.

RailDynamics.start — Method

start(t::FlatTrack)

Return 0.0 m as the beginning of the flat track.

RailDynamics.start — Method

start(t::HillyTrack)

Return start of the track in metres.