Reference

OptimalTrainControl.DavisResistance

OptimalTrainControl.finish

OptimalTrainControl.finish

OptimalTrainControl.getgrade

OptimalTrainControl.getgrade

OptimalTrainControl.getgradientacceleration

OptimalTrainControl.getgradientacceleration

OptimalTrainControl.inclinationforce

OptimalTrainControl.resistance

OptimalTrainControl.start

OptimalTrainControl.start

OptimalTrainControl.DavisResistance — Type

Empirical formula originally calculated for freight cars. The resistance (in N/kg) is given by R = a + b * v + c * v^2, where v is the vehicle speed.

OptimalTrainControl.finish — Method

finish(t::FlatTrack)

Return length of the flat track in metres.

OptimalTrainControl.finish — Method

finish(t::HillyTrack)

Return length of the track in metres.

OptimalTrainControl.getgrade — Method

getgrade(track::HillyTrack, pos)

Get the track grade (in radians) of the track at the given position pos.

OptimalTrainControl.getgrade — Method

Get the grade (in radians) of a flat track at the given position.

OptimalTrainControl.getgradientacceleration — Method

getgradientacceleration(::Scenario, position)

Return accelerating/decelerating component of the gravitational acceleration in meters per second squared [m s^-2] based on the gravitational acceleration value defined in the scenario. Positive means downhill, negative means uphill.

OptimalTrainControl.getgradientacceleration — Method

getgradientacceleration(::Track, position)

Return acceleration/decelerating component of the gravitational acceleration g = 9.81 m / s^2. Positive means downhill, negative means uphill.

OptimalTrainControl.inclinationforce — Method

Calculate the acting force component based on the current track grade. The returned value is in N/kg, i.e. force per unit mass.

OptimalTrainControl.resistance — Method

Calculate the Davis formula resistant force per unit mass.

R = a + b * v + c * v^2,

where v is the vehicle speed and a, b and c are the resistance parameters.

OptimalTrainControl.start — Method

start(t::FlatTrack)

Return 0.0 m as the beginning of the flat track.

OptimalTrainControl.start — Method

start(t::HillyTrack)

Return start of the track in metres.